Reorganise routes, index, and controllers files
