--- controllers/api/index.js
@@ -1,15 +1,0 @@
-const router = require('express').Router();
-const userRoutes = require('./userRoutes');
-const upload = require('./upload.js');
-// Using users route
-router.use('/users', userRoutes);
-
-
-// router.use('/upload', upload);
-// console.log('upload hit');
-
-router.get('/', (req, res) => {
-    res.send('Hello World!');
-});
-
-module.exports = router;
