+++ routes/userRoutes.js
@@ -1,5 +1,5 @@
 const router = require('express').Router();
-const { User } = require('../../models')
+const { User } = require('../models')
 
 // Create new user
 // Route to Sign up page
